Address PR feedback, support test network, add unsupported network
