+++ v2-components/Icons/FailedIcon/index.tsx
@@ -1,0 +1,1 @@
+export { FailedIcon } from './FailedIcon';
