+++ v2-components/Icons/GuideIcon/index.tsx
@@ -1,46 +1,1 @@
-import { Icon, IconProps } from '@chakra-ui/react';
-
-const GuideIcon = ({ width = '24px', height = '24px', color = 'white' }: IconProps) => {
-  return (
-    <Icon width={width} height={height} viewBox="0 0 24 24" fill="none" color={color}>
-      <path
-        fillRule="evenodd"
-        clipRule="evenodd"
-        d="M4.25 3.75C3.55964 3.75 3 4.30964 3 5V7C3 7.69036 3.55964 8.25 4.25 8.25H6.25C6.94036 8.25 7.5 7.69036 7.5 7V5C7.5 4.30964 6.94036 3.75 6.25 3.75H4.25ZM4.5 6.75V5.25H6V6.75H4.5Z"
-        fill="currentColor"
-      />
-      <path
-        fillRule="evenodd"
-        clipRule="evenodd"
-        d="M4.25 9.75C3.55964 9.75 3 10.3096 3 11V13C3 13.6904 3.55964 14.25 4.25 14.25H6.25C6.94036 14.25 7.5 13.6904 7.5 13V11C7.5 10.3096 6.94036 9.75 6.25 9.75H4.25ZM4.5 12.75V11.25H6V12.75H4.5Z"
-        fill="currentColor"
-      />
-      <path
-        fillRule="evenodd"
-        clipRule="evenodd"
-        d="M3 17C3 16.3096 3.55964 15.75 4.25 15.75H6.25C6.94036 15.75 7.5 16.3096 7.5 17V19C7.5 19.6904 6.94036 20.25 6.25 20.25H4.25C3.55964 20.25 3 19.6904 3 19V17ZM4.5 17.25V18.75H6V17.25H4.5Z"
-        fill="currentColor"
-      />
-      <path
-        fillRule="evenodd"
-        clipRule="evenodd"
-        d="M10.25 3.75C9.55964 3.75 9 4.30964 9 5V7C9 7.69036 9.55964 8.25 10.25 8.25H19.75C20.4404 8.25 21 7.69036 21 7V5C21 4.30964 20.4404 3.75 19.75 3.75H10.25ZM10.5 6.75V5.25H19.5V6.75H10.5Z"
-        fill="currentColor"
-      />
-      <path
-        fillRule="evenodd"
-        clipRule="evenodd"
-        d="M9 11C9 10.3096 9.55964 9.75 10.25 9.75H19.75C20.4404 9.75 21 10.3096 21 11V13C21 13.6904 20.4404 14.25 19.75 14.25H10.25C9.55964 14.25 9 13.6904 9 13V11ZM10.5 11.25V12.75H19.5V11.25H10.5Z"
-        fill="currentColor"
-      />
-      <path
-        fillRule="evenodd"
-        clipRule="evenodd"
-        d="M10.25 15.75C9.55964 15.75 9 16.3096 9 17V19C9 19.6904 9.55964 20.25 10.25 20.25H19.75C20.4404 20.25 21 19.6904 21 19V17C21 16.3096 20.4404 15.75 19.75 15.75H10.25ZM10.5 18.75V17.25H19.5V18.75H10.5Z"
-        fill="currentColor"
-      />
-    </Icon>
-  );
-};
-
-export default GuideIcon;
+export { GuideIcon } from './GuideIcon';
