+++ v2-components/Icons/KebabMenu/index.tsx
@@ -1,22 +1,1 @@
-import { Icon, IconProps } from '@chakra-ui/react';
-
-const KebabMenu = ({ width = '24px', height = '24px', color = 'white' }: IconProps) => {
-  return (
-    <Icon width={width} height={height} viewBox="0 0 24 24" fill="none" color={color}>
-      <path
-        d="M8 12C8 13.1046 7.10457 14 6 14C4.89543 14 4 13.1046 4 12C4 10.8954 4.89543 10 6 10C7.10457 10 8 10.8954 8 12Z"
-        fill="currentColor"
-      />
-      <path
-        d="M14 12C14 13.1046 13.1046 14 12 14C10.8954 14 10 13.1046 10 12C10 10.8954 10.8954 10 12 10C13.1046 10 14 10.8954 14 12Z"
-        fill="currentColor"
-      />
-      <path
-        d="M18 14C19.1046 14 20 13.1046 20 12C20 10.8954 19.1046 10 18 10C16.8954 10 16 10.8954 16 12C16 13.1046 16.8954 14 18 14Z"
-        fill="currentColor"
-      />
-    </Icon>
-  );
-};
-
-export default KebabMenu;
+export { KebabMenu } from './KebabMenu';
