+++ v2-components/Icons/SBTCIcon/SBTCIcon.tsx
@@ -1,0 +1,63 @@
+import { Icon, IconProps } from '@chakra-ui/react';
+
+interface SBTCIconProps extends IconProps {
+  width?: number;
+  height?: number;
+}
+
+export const SBTCIcon = ({ width = 34, height = 34, ...props }: SBTCIconProps) => {
+  return (
+    <Icon
+      width={width}
+      height={height}
+      viewBox="0 0 34 34"
+      fill="none"
+      sx={{
+        mask: {
+          maskType: 'alpha',
+        },
+      }}
+      {...props}
+    >
+      <mask id="mask0_1070_25650" maskUnits="userSpaceOnUse" x="0" y="0" width="34" height="34">
+        <circle cx="17" cy="17" r="17" fill="url(#paint0_linear_1070_25650)" />
+      </mask>
+      <g mask="url(#mask0_1070_25650)">
+        <circle cx="17" cy="17" r="17" fill="#31D8A4" />
+        <circle cx="16.9999" cy="16.9999" r="14.7333" fill="white" />
+      </g>
+      <circle cx="17.0001" cy="17.0001" r="13.0333" fill="#0E052F" />
+      <g clipPath="url(#clip0_1070_25650)">
+        <path
+          d="M17.0001 30.0335C24.1982 30.0335 30.0335 24.1982 30.0335 17.0001C30.0335 9.80202 24.1982 3.9668 17.0001 3.9668C9.80202 3.9668 3.9668 9.80202 3.9668 17.0001C3.9668 24.1982 9.80202 30.0335 17.0001 30.0335Z"
+          fill="#F7931A"
+        />
+        <path
+          d="M22.8561 15.3875C23.1119 13.6801 21.811 12.762 20.0336 12.1495L20.6103 9.83606L19.2027 9.48579L18.6407 11.7381C18.2709 11.6453 17.8913 11.5589 17.5125 11.4726L18.0786 9.20558L16.671 8.85449L16.0943 11.1671C15.788 11.097 15.4866 11.0286 15.195 10.9553L15.1966 10.948L13.2547 10.4633L12.8799 11.967C12.8799 11.967 13.9251 12.2065 13.9031 12.2212C14.4733 12.3637 14.5759 12.7409 14.5588 13.0406L13.9022 15.6758C13.9413 15.6856 13.9918 15.7002 14.0489 15.7222L13.8998 15.6856L12.9793 19.3773C12.9093 19.55 12.7325 19.8098 12.3334 19.7113C12.348 19.7316 11.3102 19.4563 11.3102 19.4563L10.6113 21.0675L12.4441 21.5245C12.7846 21.6101 13.1186 21.6997 13.4469 21.7836L12.8645 24.123L14.2713 24.4733L14.848 22.1599C15.2325 22.2633 15.6055 22.3595 15.9705 22.4507L15.3954 24.7543L16.803 25.1046L17.3854 22.77C19.7868 23.2246 21.5919 23.0413 22.3519 20.8696C22.9645 19.1215 22.3218 18.1122 21.0584 17.4549C21.9788 17.2431 22.6712 16.6378 22.8561 15.3875ZM19.6385 19.8986C19.2044 21.6475 16.2596 20.7018 15.305 20.4647L16.0788 17.3653C17.0335 17.6039 20.0939 18.0756 19.6385 19.8986ZM20.0743 15.3622C19.6776 16.9531 17.2274 16.1442 16.4332 15.9463L17.1337 13.1359C17.9279 13.3339 20.4882 13.7029 20.0743 15.3622Z"
+          fill="white"
+        />
+      </g>
+      <defs>
+        <linearGradient
+          id="paint0_linear_1070_25650"
+          x1="17"
+          y1="0"
+          x2="17"
+          y2="49.7121"
+          gradientUnits="userSpaceOnUse"
+        >
+          <stop stopColor="#08021E" />
+          <stop offset="1" stopColor="#1F0777" />
+        </linearGradient>
+        <clipPath id="clip0_1070_25650">
+          <rect
+            width="26.0667"
+            height="26.0667"
+            fill="white"
+            transform="translate(3.9668 3.9668)"
+          />
+        </clipPath>
+      </defs>
+    </Icon>
+  );
+};
