+++ v2-components/Icons/SettingsIcon/SettingsIcon.tsx
@@ -1,0 +1,25 @@
+import { Icon, IconProps } from '@chakra-ui/react';
+
+export const SettingsIcon = ({
+  width = '20px',
+  height = '20px',
+  color = 'whiteAlpha.700',
+  ...props
+}: IconProps) => {
+  return (
+    <Icon width={width} height={height} color={color} viewBox="0 0 20 20" fill="none" {...props}>
+      <path
+        fillRule="evenodd"
+        clipRule="evenodd"
+        d="M10 6.86666C8.26951 6.86666 6.86667 8.2695 6.86667 9.99999C6.86667 11.7305 8.26951 13.1333 10 13.1333C11.7305 13.1333 13.1333 11.7305 13.1333 9.99999C13.1333 8.2695 11.7305 6.86666 10 6.86666ZM8.11667 9.99999C8.11667 8.95986 8.95987 8.11666 10 8.11666C11.0401 8.11666 11.8833 8.95986 11.8833 9.99999C11.8833 11.0401 11.0401 11.8833 10 11.8833C8.95987 11.8833 8.11667 11.0401 8.11667 9.99999Z"
+        fill="currentColor"
+      />
+      <path
+        fillRule="evenodd"
+        clipRule="evenodd"
+        d="M9.21217 1.875L9.20592 1.87513C8.93902 1.88046 8.68486 1.99025 8.49801 2.18091C8.31118 2.37155 8.20656 2.62783 8.20658 2.89474V4.01091C7.80305 4.1309 7.41234 4.29039 7.04011 4.48706L6.27133 3.69805L6.26743 3.69408C6.17264 3.59851 6.05985 3.52264 5.93559 3.47087C5.81132 3.4191 5.67804 3.39245 5.54342 3.39245C5.40881 3.39245 5.27552 3.4191 5.15125 3.47087C5.02873 3.52192 4.91737 3.59639 4.8234 3.69008L3.69698 4.79287L3.69408 4.79572C3.59851 4.89052 3.52264 5.00331 3.47087 5.12757C3.4191 5.25184 3.39245 5.38512 3.39245 5.51974C3.39245 5.65435 3.4191 5.78764 3.47087 5.9119C3.52264 6.03617 3.59851 6.14895 3.69408 6.24375L4.48785 7.03073C4.29082 7.4034 4.13106 7.79462 4.01091 8.19868H2.89486C2.62795 8.19866 2.37155 8.30329 2.18091 8.49011C1.99025 8.67696 1.88046 8.93112 1.87513 9.19803L1.875 9.20428V10.7579C1.875 11.0283 1.98244 11.2877 2.17367 11.479C2.36491 11.6702 2.62429 11.7776 2.89474 11.7776H4.01155C4.1325 12.1784 4.29353 12.566 4.49219 12.9345L3.69408 13.7326L3.69304 13.7336C3.59796 13.8282 3.52246 13.9406 3.47087 14.0644C3.4191 14.1887 3.39245 14.322 3.39245 14.4566C3.39245 14.5912 3.4191 14.7245 3.47087 14.8487C3.52192 14.9713 3.5964 15.0827 3.6901 15.1766L4.79287 16.303L4.79572 16.3059C4.89052 16.4015 5.00331 16.4774 5.12757 16.5291C5.25183 16.5809 5.38511 16.6076 5.51974 16.6076C5.65436 16.6076 5.78764 16.5809 5.9119 16.5291C6.03617 16.4774 6.14895 16.4015 6.24375 16.3059L7.03073 15.5122C7.4034 15.7092 7.79462 15.8689 8.19868 15.9891V17.1051C8.19866 17.372 8.30329 17.6285 8.49011 17.8191C8.67697 18.0098 8.93113 18.1195 9.19803 18.1249L9.20428 18.125H10.7579C11.0283 18.125 11.2877 18.0176 11.479 17.8263C11.6702 17.6351 11.7776 17.3757 11.7776 17.1053V15.9884C12.1784 15.8675 12.566 15.7065 12.9345 15.5078L13.7326 16.3059L13.7338 16.3071C13.8283 16.4021 13.9407 16.4776 14.0644 16.5291C14.1887 16.5809 14.322 16.6076 14.4566 16.6076C14.5912 16.6076 14.7245 16.5809 14.8487 16.5291C14.9713 16.4781 15.0827 16.4036 15.1766 16.3099L16.303 15.2071L16.3059 15.2043C16.4015 15.1095 16.4774 14.9967 16.5291 14.8724C16.5809 14.7482 16.6076 14.6149 16.6076 14.4803C16.6076 14.3456 16.5809 14.2124 16.5291 14.0881C16.4776 13.9644 16.4021 13.852 16.3071 13.7575L16.3059 13.7562L15.5078 12.9581C15.7039 12.5945 15.8633 12.2123 15.9837 11.8171H17.1053C17.3757 11.8171 17.6351 11.7097 17.8263 11.5184C18.0176 11.3272 18.125 11.0678 18.125 10.7974V9.21217L18.1249 9.20592C18.1195 8.93903 18.0098 8.68486 17.8191 8.49801C17.6285 8.31118 17.3722 8.20656 17.1053 8.20658H15.9891C15.8691 7.80305 15.7096 7.41234 15.5129 7.04011L16.302 6.27133L16.3059 6.26743C16.4015 6.17264 16.4774 6.05985 16.5291 5.93559C16.5809 5.81133 16.6076 5.67804 16.6076 5.54342C16.6076 5.4088 16.5809 5.27551 16.5291 5.15125C16.4784 5.02959 16.4047 4.91893 16.3119 4.82538L15.226 3.7002L15.2201 3.69408C15.1253 3.59851 15.0125 3.52265 14.8882 3.47087C14.764 3.4191 14.6307 3.39245 14.4961 3.39245C14.3614 3.39245 14.2282 3.4191 14.1039 3.47087C13.9801 3.52245 13.8677 3.59795 13.7731 3.69304L13.772 3.69408L12.9739 4.49219C12.6055 4.29353 12.2179 4.1325 11.8171 4.01155V2.89474C11.8171 2.62428 11.7097 2.36491 11.5184 2.17367C11.3272 1.98244 11.0678 1.875 10.7974 1.875H9.21217ZM9.45658 4.18947V3.125H10.5671V4.18947L10.5671 4.19622C10.5696 4.4222 10.647 4.64097 10.7873 4.81817C10.9256 4.9929 11.1173 5.11743 11.333 5.17286C11.7433 5.28526 12.1376 5.44978 12.5061 5.66241L12.5092 5.66418C12.7029 5.7745 12.9275 5.81828 13.1485 5.78882C13.3695 5.75935 13.5748 5.65826 13.7329 5.50102L13.734 5.49984L14.4931 4.74078L15.2629 5.53848L14.5059 6.27604L14.5052 6.27677C14.3407 6.43754 14.2365 6.64989 14.2098 6.87832C14.184 7.09995 14.2327 7.32369 14.348 7.51428C14.5554 7.88473 14.7162 8.2794 14.8268 8.68931C14.882 8.90554 15.0067 9.09776 15.1818 9.23639C15.359 9.37666 15.5778 9.4541 15.8038 9.45654L15.8105 9.45658H16.875V10.5671H15.8105C15.5899 10.5671 15.3752 10.6387 15.1987 10.7711C15.0222 10.9034 14.8934 11.0895 14.8316 11.3013L14.8286 11.3119C14.7162 11.7241 14.5511 12.1202 14.3376 12.4903L14.3358 12.4934C14.2255 12.6871 14.1817 12.9117 14.2112 13.1327C14.2407 13.3537 14.3417 13.559 14.499 13.7171L14.5002 13.7183L15.2604 14.4785L14.4583 15.2639L13.6946 14.5002L13.6934 14.499C13.5353 14.3417 13.33 14.2407 13.109 14.2112C12.888 14.1817 12.6635 14.2255 12.4697 14.3358L12.4666 14.3376C12.0981 14.5502 11.7038 14.7147 11.2935 14.8271C11.0778 14.8826 10.8861 15.0071 10.7478 15.1818C10.6075 15.359 10.5301 15.5778 10.5277 15.8038L10.5276 15.8105V16.875H9.44868V15.8105L9.44865 15.8038C9.44621 15.5778 9.36877 15.359 9.22849 15.1818C9.08987 15.0067 8.89765 14.882 8.68141 14.8268C8.2715 14.7162 7.87684 14.5554 7.50638 14.348C7.31579 14.2327 7.09206 14.184 6.87042 14.2098C6.64294 14.2364 6.4314 14.3399 6.27087 14.5032L5.52078 15.2597L4.73614 14.4583L5.49984 13.6946L5.50102 13.6934C5.65826 13.5353 5.75935 13.33 5.78882 13.109C5.81828 12.888 5.7745 12.6635 5.66418 12.4697L5.66241 12.4666C5.44978 12.0981 5.28526 11.7038 5.17286 11.2935C5.11743 11.0779 4.9929 10.8862 4.81817 10.7478C4.64098 10.6075 4.4222 10.5301 4.19622 10.5277L4.18947 10.5276H3.125V9.44868H4.18947L4.19622 9.44865C4.4222 9.44621 4.64098 9.36877 4.81817 9.22849C4.99327 9.08987 5.11797 8.89764 5.17321 8.6814C5.28381 8.2715 5.44462 7.87684 5.652 7.50639C5.76729 7.3158 5.81603 7.09206 5.79017 6.87042C5.76363 6.64291 5.66009 6.43134 5.49677 6.2708L4.74028 5.52078L5.53959 4.73823L6.27677 5.49481C6.43754 5.65925 6.64989 5.76352 6.87832 5.79017C7.09996 5.81603 7.32369 5.76729 7.51429 5.652C7.88473 5.44462 8.27939 5.28381 8.6893 5.17321C8.90554 5.11797 9.09776 4.99327 9.23639 4.81817C9.37666 4.64098 9.4541 4.4222 9.45654 4.19622L9.45658 4.18947Z"
+        fill="currentColor"
+      />
+    </Icon>
+  );
+};
