+++ v2-components/Icons/StakingIcon/index.tsx
@@ -1,0 +1,1 @@
+export { StakingIcon } from './StakingIcon';
