+++ v2-components/StakingLogo/index.tsx
@@ -1,39 +1,1 @@
-import { Icon, IconProps, useBreakpointValue } from '@chakra-ui/react';
-
-interface StakingLogoProps extends IconProps {
-  color?: string;
-  textColor?: 'white';
-}
-
-const StakingLogo = ({ color = '#00D1FF', textColor = 'white' }: StakingLogoProps) => {
-  const size = useBreakpointValue({
-    base: 'small',
-    lg: 'large',
-  });
-
-  if (size === 'small')
-    return (
-      <Icon width="29" height="21" viewBox="0 0 29 21" fill="none">
-        <path
-          d="M6.81239 5.45766C6.61525 5.22805 6.37254 5.11297 6.08372 5.11297H0.177C0.124536 5.11297 0.0816109 5.09606 0.0492845 5.06224C0.0164282 5.02898 0 4.99134 0 4.95098V0.89807C0 0.85771 0.0164282 0.820622 0.0492845 0.786807C0.0816109 0.752992 0.124536 0.736084 0.177 0.736084H6.41864C7.99362 0.736084 9.35239 1.39166 10.4944 2.70173L12.0106 4.58611L9.05721 8.25396L6.81239 5.45766ZM17.7207 2.68155C18.8627 1.38457 20.2278 0.736084 21.8161 0.736084H28.0381C28.0906 0.736084 28.1298 0.749719 28.1563 0.776444C28.1822 0.803714 28.1955 0.844074 28.1955 0.89807V4.95098C28.1955 4.99134 28.1822 5.02898 28.1563 5.06224C28.1298 5.09606 28.0906 5.11297 28.0381 5.11297H22.1314C21.8426 5.11297 21.5998 5.22805 21.4027 5.45766L17.0514 10.8479L21.4223 16.2785C21.6195 16.495 21.8553 16.603 22.1314 16.603H28.0381C28.0906 16.603 28.1298 16.6199 28.1563 16.6538C28.1822 16.6876 28.1955 16.7318 28.1955 16.7852V20.8381C28.1955 20.8785 28.1822 20.9161 28.1563 20.9494C28.1298 20.9832 28.0906 21.0001 28.0381 21.0001H21.8161C20.2278 21.0001 18.8691 20.3451 17.7403 19.0345L14.1176 14.5359L10.4944 19.0345C9.35239 20.3451 7.98727 21.0001 6.39903 21.0001H0.177C0.124536 21.0001 0.0847905 20.9832 0.0588234 20.9494C0.0323264 20.9156 0.0196078 20.8719 0.0196078 20.8174V16.7645C0.0196078 16.7241 0.0323264 16.687 0.0588234 16.6532C0.0847905 16.6194 0.124536 16.6025 0.177 16.6025H6.08372C6.35929 16.6025 6.602 16.488 6.81239 16.2578L11.0848 10.9488L17.7207 2.68155Z"
-          fill={color}
-        />
-      </Icon>
-    );
-
-  return (
-    <Icon width="116" height="21" viewBox="0 0 116 21" fill="none">
-      <path
-        d="M47.9982 5.14C48.6982 5.14 49.2582 5.7 49.2582 6.4V7.08H51.7782V6.12C51.7782 4.42 50.2982 3 48.6382 3H42.4982C40.7782 3 39.3782 4.44 39.3782 6.12V7.84C39.3782 9.56 40.8382 10.94 42.4982 10.94H48.2182C48.9182 10.94 49.4782 11.52 49.4782 12.22V13.6C49.4782 14.3 48.9182 14.88 48.2182 14.88H42.9982C42.2982 14.88 41.7382 14.3 41.7382 13.6V12.94H39.2182V13.9C39.2182 15.6 40.6982 17 42.3582 17H48.8582C50.5982 17 51.9782 15.56 51.9782 13.9V11.92C51.9782 10.2 50.5382 8.82 48.8582 8.82H43.1582C42.4582 8.82 41.8782 8.24 41.8782 7.54V6.42C41.8782 5.72 42.4582 5.14 43.1582 5.14H47.9982ZM60.901 15H58.741C58.081 15 57.641 14.54 57.641 13.88V8.18H60.921V6.16H57.641V3.78H55.321V6.16H53.441V8.18H55.321V14.22C55.321 15.8 56.581 17 58.121 17H60.901V15ZM66.0376 15C65.3976 15 64.8776 14.5 64.8776 13.84V9.32C64.8776 8.68 65.3976 8.18 66.0376 8.18H70.3576V14.3C70.3576 14.7 70.0376 15 69.6576 15H66.0376ZM68.6176 17L70.3576 15.52V17H72.6576V6.16H65.4176C63.8976 6.16 62.5776 7.42 62.5776 9V14.16C62.5776 15.68 63.8376 17 65.4176 17H68.6176ZM82.639 6.16L77.559 11.02V2.28H75.259V17H77.559V11.86L83.099 17H86.119L79.859 11.36L85.619 6.16H82.639ZM87.6813 6.16V17H90.0013V6.16H87.6813ZM87.6813 2.44V4.76H90.0013V2.44H87.6813ZM94.9223 8.86C94.9223 8.46 95.2423 8.18 95.6223 8.18H99.4823C100.122 8.18 100.622 8.68 100.622 9.32V17H102.942V9C102.942 7.42 101.642 6.16 100.102 6.16H96.6023L94.9223 7.58V6.16H92.6223V17H94.9223V8.86ZM108.772 14.66C108.132 14.66 107.612 14.16 107.612 13.52V9.32C107.612 8.68 108.132 8.18 108.772 8.18H113.252V13.98C113.252 14.38 112.932 14.66 112.552 14.66H108.772ZM112.732 20.92C114.312 20.92 115.572 19.62 115.572 18.1V6.16H108.152C106.632 6.16 105.312 7.42 105.312 9V13.84C105.312 15.36 106.572 16.68 108.152 16.68H111.512L113.252 15.2V17.84C113.252 18.5 112.752 19 112.092 19H109.272C108.632 19 108.112 18.5 108.112 17.84V17.5H105.852V18.1C105.852 19.68 107.152 20.92 108.692 20.92H112.732Z"
-        fill={textColor}
-      />
-
-      <path
-        d="M6.81239 4.59862C6.61525 4.37499 6.37254 4.2629 6.08372 4.2629H0.177C0.124536 4.2629 0.0816109 4.24643 0.0492845 4.2135C0.0164282 4.1811 0 4.14444 0 4.10514V0.157767C0 0.118458 0.0164282 0.0823364 0.0492845 0.0494019C0.0816109 0.0164673 0.124536 0 0.177 0H6.41864C7.99362 0 9.35239 0.638506 10.4944 1.91445L12.0106 3.74976L9.05722 7.3221L6.81239 4.59862ZM17.7207 1.8948C18.8627 0.6316 20.2278 0 21.8161 0H28.0381C28.0906 0 28.1298 0.0132801 28.1563 0.039309C28.1822 0.0658691 28.1955 0.105178 28.1955 0.157767V4.10514C28.1955 4.14444 28.1822 4.1811 28.1563 4.2135C28.1298 4.24643 28.0906 4.2629 28.0381 4.2629H22.1314C21.8426 4.2629 21.5998 4.37499 21.4027 4.59862L17.0514 9.8485L21.4223 15.1377C21.6195 15.3486 21.8553 15.4538 22.1314 15.4538H28.0381C28.0906 15.4538 28.1298 15.4702 28.1563 15.5032C28.1822 15.5361 28.1955 15.5791 28.1955 15.6312V19.5785C28.1955 19.6179 28.1822 19.6545 28.1563 19.6869C28.1298 19.7198 28.0906 19.7363 28.0381 19.7363H21.8161C20.2278 19.7363 18.8691 19.0983 17.7403 17.8219L14.1176 13.4405L10.4944 17.8219C9.35239 19.0983 7.98726 19.7363 6.39903 19.7363H0.177C0.124536 19.7363 0.0847905 19.7198 0.0588234 19.6869C0.0323264 19.654 0.0196078 19.6115 0.0196078 19.5584V15.611C0.0196078 15.5717 0.0323264 15.5356 0.0588234 15.5026C0.0847905 15.4697 0.124536 15.4532 0.177 15.4532H6.08372C6.35929 15.4532 6.602 15.3417 6.81239 15.1175L11.0848 9.94677L17.7207 1.8948Z"
-        fill={color}
-      />
-    </Icon>
-  );
-};
-
-export default StakingLogo;
+export { StakingLogo } from './StakingLogo';
